TASK: Convert Drawer/MenuItem component to typescript
