+++ packages/neos-ui/src/Containers/Drawer/MenuItem/index.ts
@@ -1,0 +1,10 @@
+/*
+ * This file is part of the Neos.Neos.Ui package.
+ *
+ * (c) Contributors of the Neos Project - www.neos.io
+ *
+ * This package is Open Source Software. For the full copyright and license
+ * information, please view the LICENSE file which was distributed with this
+ * source code.
+ */
+export {MenuItem as default} from './MenuItem';
